EntryElement text field is correctly position on iPad. Closes #1.
The issue is that the default width for a new UITableViewCell is 320,
so if the FlexibleLeftMargin flag is set the cell is resized on the
UITableView and the field is essentially shifted to the right as the
cell’s width is increased rather than expanding as the cell’s width is
increased. Of course, when a cell is de-queued the cell already has the
correct width of 768 (not 320) and therefore the field is correctly
positioned. Therefore, this issue should only affect iPad. I have only
tested on iPad not iPhone.