EntryElement UITextField adjusted Y alignment and font size. Closes #2.
Uses the UITableViewCell.TextLabel.Font for the UITextField in order to
create a consistent user interface. This aspect negates the need for a
-1 Y adjustment that was made so that the UITextField is correctly
aligned with the base line of the TextLabel.

Added a right margin between the entry field and the right of the cell.

May not work on iPhone.